refactor tests to conform to PEP8 standards
